feat: added seeders and fatories

+++ backend/database/migrations/2024_12_05_114506_create_student_courses_table.php
@@ -8,16 +8,19 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('student_courses', function (Blueprint $table) {
-            $table->id();
-            $table->bigInteger('student_id')->unsigned();
-            $table->bigInteger('course_id')->unsigned();
 
-            $table->foreign('student_id')->references('id')->on('students')->onDelete('cascade');
-            $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
+        if (!Schema::hasTable('student_courses')) {
+            Schema::create('student_courses', function (Blueprint $table) {
+                $table->id();
+                $table->bigInteger('student_id')->unsigned();
+                $table->bigInteger('course_id')->unsigned();
 
-            $table->timestamps();
-        });
+                $table->foreign('student_id')->references('id')->on('students')->onDelete('cascade');
+                $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
+                $table->unique(['student_id', 'course_id']);
+                $table->timestamps();
+            });
+        }
     }
